+++ simplex/test.cpp
@@ -37,16 +37,16 @@ testResultCodes allTests::runAllTests () {
 }
 
 testErrorCodes allTests::simplTest() {
-	float* correctxValues = new float[NUMVAR];
+	double* correctxValues = new double[NUMVAR];
 	correctxValues[0] = 10.0;
 	correctxValues[1] = 5.0;
-	float correctOptimalValue = 2750.0;
+	double correctOptimalValue = 2750.0;
 	testErrorCodes result = TEST_SUCCESS;
 
 	simpl* test = new simpl(NUMVAR, NUMCONSTR);
 	int variableIter;
-	float optimalValue = -8.2;
-	float* testxVal = new float[NUMVAR];
+	double optimalValue = -8.2;
+	double* testxVal = new double[NUMVAR];
 	test->changeValue(1.0, 1, 1);
 	test->changeValue(2.0, 1, 2);
 	test->changeValue(20.0, 1, 3);
@@ -115,8 +115,8 @@ testErrorCodes allTests::negrowTest() {
 
 testErrorCodes allTests::unboundedTest() {
 	simpl* test = new simpl(2, 2);
-	float* xRes = new float[2];
-	float optVal = -1;
+	double* xRes = new double[2];
+	double optVal = -1;
 	testErrorCodes result = TEST_SUCCESS;
 	test->changeValue(1.0, 1, 1);
 	test->changeValue(-2.0, 1, 2);
@@ -138,8 +138,8 @@ testErrorCodes allTests::unboundedTest() {
 
 testErrorCodes allTests::infeasibleTest() {
 	simpl* test = new simpl(2, 2);
-	float* xRes = new float[2];
-	float optVal = -1;
+	double* xRes = new double[2];
+	double optVal = -1;
 	testErrorCodes result = TEST_SUCCESS;
 	test->changeValue(-1.0, 1, 1);
 	test->changeValue(2.0, 1, 2);
